refactor: seperate component logics for reinfore reuseable
