+++ src/shared/ui/hooks/useToasts.js
@@ -1,0 +1,24 @@
+import { useState } from 'react';
+
+export const useToasts = () => {
+  const [toasts, setToasts] = useState([]);
+
+  const showToast = (message, type = 'success', duration = 3000) => {
+    const id = Date.now();
+    const newToast = {
+      id,
+      message,
+      type,
+      duration,
+      isVisible: true,
+      onClose: () => removeToast(id),
+    };
+    setToasts(prev => [...prev, newToast]);
+  };
+
+  const removeToast = (id) => {
+    setToasts(prev => prev.filter(toast => toast.id !== id));
+  };
+
+  return { toasts, showToast };
+};
